Testing kfree

+++ libc/free.c
@@ -2,5 +2,5 @@
 #include <unistd.h>
 
 void free(void *ptr) {
-    syscall2(SYSCALL_FREE, ptr);
+    syscall2(SYSCALL_FREE, (uint64_t) ptr);
 }
